+++ tusker/config.py
@@ -21,9 +21,15 @@ class Config:
         self.schema = SchemaConfig(data['schema'])
         self.migrations = MigrationsConfig(data['migrations'])
         self.database = DatabaseConfig(data['database'])
+        self.migra = MigraConfig(data['migra'])
 
     def __str__(self):
-        return 'Config(schema={}, migrations={}, database={})'.format(self.schema, self.migrations, self.database)
+        return 'Config(schema={}, migrations={}, database={}, migra={})'.format(
+            self.schema,
+            self.migrations,
+            self.database,
+            self.migra
+        )
 
 
 class ConfigReader:
@@ -32,12 +38,12 @@ class ConfigReader:
         self.data = data
         self.path = path
 
-    def get(self, name, type, required=False):
+    def get(self, name, type, required=False, default=None):
         if name not in self.data:
             if required:
-                raise ConfigError.missing('{}.{}'.format(path, name))
+                raise ConfigError.missing('{}.{}'.format(self.path, name))
             else:
-                return None
+                return default
         value = self.data[name]
         if not isinstance(value, type):
             raise ConfigError.invalid(name, 'Not of type {}'.format(type))
@@ -100,6 +106,13 @@ class DatabaseConfig:
             args['dbname'] = 'tusker'
         args.update(override)
         return args
+
+
+class MigraConfig:
+    def __init__(self, data):
+        data = ConfigReader(data, 'migra')
+        self.safe = data.get('safe', bool, default=False)
+        self.privileges = data.get('with_privileges', bool, default=False)
 
 
 class ConfigError(RuntimeError):